refactor: align the Next.js SSR example naming with Svelte
Merge the client-only and hydrate wrappers into one TodoPanel that takes an optional snapshot (ssr={!!snapshot}), sharing TodoList and TodoForm — so a component only knows whether a snapshot was passed, not whether it's for SSR. The server prefetch keeps its own component (PrefetchedTodoPanel) since it has to be a server component; the hydrate column now writes through the client TodoForm, with the Server column still demonstrating server-action writes. Updates the recipe's React tab to the same TodoPanel/TodoList vocabulary as the Svelte tab.
